requirements.txt
flask
pandas
matplotlib
seaborn
google-cloud-aiplatform
python-dotenv
